# packages/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: NEBULA Control API
  version: "1.0.0-phase2"
  summary: Model registry, credentials and deployment lifecycle.
  description: |
    The control plane is the only writer of desired state. Everything here is
    org-scoped by the caller's credential: `org_id` is never a parameter, in a
    path, a query or a body.

    **Deployments are reconciled intent.** A successful `POST /v1/deployments`
    returns `202 Accepted` and records desired state in PostgreSQL. The Kubernetes
    controller that acts on it arrives in Phase 5, so in this build
    `status.reconciled` is always `false` and `status.observed_generation` is
    always `0`. Responses that describe unreconciled state carry a `note` saying
    so, rather than leaving a caller to infer that work is underway.
  license:
    name: Apache-2.0
  contact:
    name: NEBULA
    url: https://github.com/adityasatwar321/nebula
servers:
  - url: /
    description: The service itself, reached through the gateway in a deployed cluster.
security:
  - apiKey: []
...
paths:
...
  /v1/models:
    post:
      tags: [registry]
      operationId: createModel
      summary: Register a model family.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateModel" }
      responses:
        "201":
          description: The created model.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Model" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/models/registry:
    get:
      tags: [registry]
      operationId: listModels
      summary: List registered models.
      description: |
        Distinct from the OpenAI-compatible `GET /v1/models`, which lists routes in
        OpenAI's shape and is served by the gateway. This path returns NEBULA's own
        model records.
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: A page of models.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Model" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/models/{model_id}:
    parameters:
      - name: model_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [registry]
      operationId: getModel
      summary: Fetch a model.
      responses:
        "200":
          description: The model.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Model" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [registry]
      operationId: updateModel
      summary: Change a model's mutable metadata.
      description: |
        Only `family` and `description`. The name is immutable: it appears in
        Kubernetes object names and in every audit record referencing the model, so
        renaming it would make history unreadable. A model that needs a different
        name is a new model.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                family: { type: [string, "null"] }
                description: { type: [string, "null"] }
      responses:
        "200":
          description: The updated model.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Model" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [registry]
      operationId: deleteModel
      summary: Soft-delete a model.
      description: Refused while any of its versions backs a live deployment.
      responses:
        "204": { description: Deleted. }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: nbk
      description: |
        `Authorization: Bearer nbk_<key>`. Session tokens share this header in the
        deployed system but are not accepted by this service until the dashboard
        arrives in Phase 7; a non-`nbk_` bearer is refused with code
        `unsupported_credential` rather than a generic 401, because "your
        credential type is not supported yet" and "your key is wrong" are different
        problems.
  parameters:
    limit:
      name: limit
      in: query
      description: Page size, 1-200. Defaults to 50.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    cursor:
      name: cursor
      in: query
      description: |
        The `next_cursor` from a previous page. Cursor-based, not offset-based:
        offset paging skips and duplicates rows while the collection is being
        written to.
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: The request is malformed or semantically invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: |
        No usable credential. The message never distinguishes "no such key" from
        "wrong key": that difference is an oracle for enumerating valid prefixes.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: The credential lacks the required scope or role.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: |
        No such resource. Also returned for a resource in another organization, so
        existence elsewhere is never disclosed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: A uniqueness or state precondition failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      description: |
        One shape everywhere, OpenAI-compatible so existing SDK error handling
        works. It carries nothing internal: no pod names, no SQL detail, no
        upstream bodies. Those belong in logs and traces, which are operator-only.
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message: { type: string }
            type:
              type: string
              enum:
                - invalid_request_error
                - authentication_error
                - permission_error
                - not_found_error
                - conflict_error
                - unprocessable_entity
                - rate_limit_error
                - internal_error
                - upstream_error
                - service_unavailable
                - timeout_error
            code: { type: string }
            param: { type: string }
            request_id: { type: string }
    Page:
      type: object
      required: [data, next_cursor, has_more]
      properties:
        data: { type: array, items: {} }
        next_cursor:
          type: [string, "null"]
          description: Absent when this is the last page.
        has_more: { type: boolean }
...
    ModelTask:
      type: string
      enum: [chat, completion, embedding]
    Model:
      type: object
      required: [id, name, task, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name:
          type: string
          pattern: "^[a-z0-9]([a-z0-9._-]{0,62}[a-z0-9])?$"
          description: Immutable once created.
        family: { type: [string, "null"] }
        task: { $ref: "#/components/schemas/ModelTask" }
        description: { type: [string, "null"] }
        created_by: { type: [string, "null"], format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateModel:
      type: object
      required: [name, task]
      properties:
        name: { type: string, pattern: "^[a-z0-9]([a-z0-9._-]{0,62}[a-z0-9])?$" }
        task: { $ref: "#/components/schemas/ModelTask" }
        family: { type: [string, "null"] }
        description: { type: [string, "null"] }
...